Avoid reading unneeded data when server-side GET receives SEEK from client.
This approach uses an "unread" operation to support peeking at the
next pending control-packet, and unreading if it turns out not to be a
SEEK.  The unread is stored in the handle, so this means passing a
handle through lots of functions that formerly took an fd.  Not that
pretty.  Should probably have just used recv(... MSG_PEEK), which it
turns out is supported by rsockets.  But the current approach does
work.